Refactor PushService to separate subscriber logic.
